requirements.txt
python-dotenv==1.0.0
supabase==1.0.4
streamlit==1.27.0
